Int test refactor: extract helpful classes.
