+++ jvm/workbookapp/src/integration-test/kotlin/integrationtest/rcimport/RowCount.kt
@@ -1,0 +1,9 @@
+package integrationtest.rcimport
+
+import org.wycliffeassociates.otter.common.data.model.ContentType
+
+data class RowCount(
+    val collections: Int,
+    val links: Int,
+    val contents: Map<ContentType, Int>
+)
